Provide the option to pass in an enviroment
By default, when using loadstring, it only has access to global values. If you want it to have access to local values (upvalues), then you *could* use one of the various debug functions, but it adds a lot of extra overhead, as every time the function is called, it needs to build it's own representation of what the environment looks like. This iterating, building tables, and calling debug functions is pretty excessive when all you want to do is get access to one or two upvalues, so instead we provide the env table. Similiar to how the fformat function worked, you can pass in values directly, but instead of them being numeric values, they are key/value pairs, and since you'll know what values you need access to, you can only bother passing in what's relevant to the string, rather than every single upvalue every time.

Due to the usage of setfenv, TL is now limited to either LuaJIT or Lua 5.1 or below.