Add recursion system, and env system
Upvalues do not exist in the function by default, and rather than iterate through all upvalues for potentially only one access every time TL is called, we make it so you can simply pass in a custom env table. the env table will add a mt to it that has a __index pointing to _G so we can still access global values if local ones are added.

We also add a hidden third param called recursive, that is only set to true internally. This is so that when calling something like TL"%{'Hello '.. %{'wor' .. 'ld'}}", rather than world getting returned as just `world`, and causing the second concat to fail because there's no `world` variable, we instead replace it with an id pointing towards the actual value that was returned, and place that in env. Only if recursive is false do we return the true string value as expected.